+++ engine/engine/iterative_deepening.h
@@ -1,4 +1,4 @@
 #pragma once
 #include "game.h"
 
-int TimeSearch(Game& game, std::chrono::steady_clock::time_point deadline, int* reachedDepth = nullptr);
+float TimeSearch(Game& game, std::chrono::steady_clock::time_point deadline, int* reachedDepth = nullptr);
