+++ engine/engine/minimax.h
@@ -4,4 +4,4 @@
 
 extern int totalNodesSearched;
 
-int Minimax(Game& game, const unsigned int depth, int alpha = -0x10000, int beta = 0x10000, int parentId = 0, bool visualize = 0, std::optional<std::chrono::steady_clock::time_point> deadline = std::nullopt); //returns score of position relative to the side to move
+float Minimax(Game& game, const unsigned int depth, float alpha = -0x10000, float beta = 0x10000, int parentId = 0, bool visualize = 0, std::optional<std::chrono::steady_clock::time_point> deadline = std::nullopt); //returns score of position relative to the side to move
